Get TS up and running
https://svelte.dev/blog/svelte-and-typescript
