+++ package.json
@@ -14,19 +14,20 @@
 		"@sveltejs/adapter-auto": "next",
 		"@sveltejs/adapter-netlify": "^1.0.0-next.51",
 		"@sveltejs/kit": "next",
+		"@tsconfig/svelte": "^3.0.0",
 		"@typescript-eslint/eslint-plugin": "^5.10.1",
 		"@typescript-eslint/parser": "^5.10.1",
 		"eslint": "^7.32.0",
 		"eslint-config-prettier": "^8.3.0",
 		"eslint-plugin-svelte3": "^3.2.1",
+		"mdsvex": "^0.10.6",
 		"prettier": "^2.5.1",
 		"prettier-plugin-svelte": "^2.5.0",
 		"svelte": "^3.44.0",
-		"svelte-check": "^2.2.6",
-		"svelte-preprocess": "^4.10.1",
+		"svelte-check": "^2.10.3",
+		"svelte-preprocess": "^4.10.7",
 		"tslib": "^2.3.1",
-		"typescript": "~4.6.2",
-		"mdsvex": "^0.10.6"
+		"typescript": "~4.6.2"
 	},
 	"type": "module",
 	"dependencies": {
