+++ src/data/interfaces/interfaces.ts
@@ -1,0 +1,10 @@
+export type OutlineObject = {
+	letterGroupings?: string[];
+	specialOutlineMeanings: string[];
+	lines: LineDetails[];
+};
+
+type LineDetails = {
+	path: string;
+	translateValues: [number, number];
+};
